Split serializers apart into multiple files to mirror the models
closes: #7072
https://pulp.plan.io/issues/7072
[nocoverage]
