+++ CHANGES/7072.misc
@@ -1,0 +1,1 @@
+Split serializer.py into many files mirroring how the models are split up.
