+++ pulp_rpm/app/serializers/__init__.py
@@ -1,0 +1,29 @@
+from .advisory import (  # noqa
+    MinimalUpdateRecordSerializer,
+    UpdateCollectionSerializer,
+    UpdateRecordSerializer,
+)
+from .comps import (  # noqa
+    PackageCategorySerializer,
+    PackageEnvironmentSerializer,
+    PackageGroupSerializer,
+    PackageLangpacksSerializer,
+)
+from .custom_metadata import RepoMetadataFileSerializer  # noqa
+from .distribution import (  # noqa
+    AddonSerializer,
+    ChecksumSerializer,
+    DistributionTreeSerializer,
+    ImageSerializer,
+    VariantSerializer,
+)
+from .modulemd import ModulemdSerializer, ModulemdDefaultsSerializer  # noqa
+from .package import PackageSerializer, MinimalPackageSerializer  # noqa
+from .repository import (  # noqa
+    CopySerializer,
+    RpmDistributionSerializer,
+    RpmPublicationSerializer,
+    RpmRemoteSerializer,
+    RpmRepositorySerializer,
+    RpmRepositorySyncURLSerializer,
+)
