+++ pulp_rpm/app/serializers/custom_metadata.py
@@ -1,0 +1,30 @@
+from gettext import gettext as _
+from rest_framework import serializers
+
+from pulpcore.plugin.serializers import (
+    ContentChecksumSerializer,
+    SingleArtifactContentUploadSerializer,
+)
+
+from pulp_rpm.app.models import RepoMetadataFile
+
+
+class RepoMetadataFileSerializer(
+    SingleArtifactContentUploadSerializer, ContentChecksumSerializer
+):
+    """
+    RepoMetadataFile serializer.
+    """
+
+    data_type = serializers.CharField(help_text=_("Metadata type."))
+    checksum_type = serializers.CharField(help_text=_("Checksum type for the file."))
+    checksum = serializers.CharField(help_text=_("Checksum for the file."))
+
+    class Meta:
+        fields = SingleArtifactContentUploadSerializer.Meta.fields + (
+            "data_type",
+            "checksum_type",
+            "checksum",
+            "sha256",
+        )
+        model = RepoMetadataFile
